Check that no element is used more than produced
That is, except for protons, which are just used.
This way, for every 7Li used in (p+7Li), one 7Li should have been
produced by (e+7Be), and so on. For the use of 3He, we try to
scale the usage down so that PPI and PPII doesn't use more 3He
than what was produced by p+p.

Per now this scaling shows that PPII happens much more than PPI.
This should be PPI=69% and PPII=31%, so some bug or error exists.

Possibly related to #1.