Fix convection test
The test was to strict, it was set to get exactly the
values from the sanity test. This should only
be the case for mu=0.6. I have mu=0.618, which is close,
so I will test that values are close to the sanity
test, but not require exactly the right result.